[#424] Refine AdjustIncomeTarget as a component
