+++ frontend/src/pages/cases/components/index.js
@@ -11,3 +11,4 @@ export { default as IncomeDriversDropdown } from "./IncomeDriversDropdown";
 export { default as BinningDriverForm } from "./BinningDriverForm";
 export { default as ScenarioModelingForm } from "./ScenarioModelingForm";
 export { default as AllDriverTreeSelector } from "./AllDriverTreeSelector";
+export { default as AdjustIncomeTarget } from "./AdjustIncomeTarget";
