+++ frontend/src/pages/cases/steps/AssessImpactMitigationStrategies.js
@@ -12,20 +12,7 @@ import {
   CaseVisualState,
   CaseUIState,
 } from "../store";
-import {
-  Row,
-  Col,
-  Card,
-  Space,
-  Carousel,
-  Form,
-  Button,
-  Select,
-  Switch,
-  message,
-  Modal,
-  InputNumber,
-} from "antd";
+import { Row, Col, Card, Space, Carousel, Form, Button, message } from "antd";
 import {
   ChartBiggestImpactOnIncome,
   ChartMonetaryImpactOnIncome,
@@ -35,17 +22,12 @@ import {
 import {
   BinningDriverForm,
   SegmentSelector,
-  AllDriverTreeSelector,
+  AdjustIncomeTarget,
 } from "../components";
 import { map, groupBy, isEqual, isEmpty } from "lodash";
 import { commodities } from "../../../store/static";
 import { api } from "../../../lib";
-import {
-  removeUndefinedObjectValue,
-  InputNumberThousandFormatter,
-  determineDecimalRound,
-} from "../../../lib";
-import { thousandFormatter } from "../../../components/chart/options/common";
+import { removeUndefinedObjectValue } from "../../../lib";
 import { LeftOutlined, RightOutlined } from "@ant-design/icons";
 
 /**
@@ -64,9 +46,6 @@ const AssessImpactMitigationStrategies = ({
   const { enableEditCase } = CaseUIState.useState((s) => s.general);
 
   const [selectedSegment, setSelectedSegment] = useState(null);
-  const [showAdjustIncomeModal, setShowAdjustIncomeModal] = useState(false);
-  const [percentageSensitivity, setPercentageSensitivity] = useState(true);
-  const [adjustedValues, setAdjustedValues] = useState({});
 
   const carouselChartRef = useRef(null);
 
@@ -78,43 +57,6 @@ const AssessImpactMitigationStrategies = ({
       caseButton: value,
     }));
   };
-
-  // initial load adjusted income target
-  useEffect(() => {
-    if (!isEmpty(sensitivityAnalysis?.config)) {
-      setAdjustedValues(
-        Object.keys(sensitivityAnalysis?.config)
-          .filter((x) => x.includes("adjusted-target"))
-          .map((x) => ({ [x]: sensitivityAnalysis.config[x] }))
-          .reduce((a, c) => ({ ...a, ...c }), {})
-      );
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  const adustedTargetChange = useMemo(() => {
-    if (percentageSensitivity) {
-      const res =
-        adjustedValues?.[
-          `${selectedSegment}_absolute-increase_adjusted-target`
-        ] || 0;
-      return thousandFormatter(res, 2);
-    }
-    const res =
-      adjustedValues?.[
-        `${selectedSegment}_percentage-increase_adjusted-target`
-      ] || 0;
-    return `${res}%`;
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [selectedSegment, percentageSensitivity, adjustedValues]);
-
-  const currentSegmentDetail = useMemo(() => {
-    if (selectedSegment && currentCase?.segments?.length) {
-      const res = currentCase.segments.find((s) => s.id === selectedSegment);
-      return res;
-    }
-    return null;
-  }, [currentCase?.segments, selectedSegment]);
 
   const handleSaveVisualization = useCallback(() => {
     if (!enableEditCase) {
@@ -179,56 +121,6 @@ const AssessImpactMitigationStrategies = ({
     prevSensitivityAnalysis,
     sensitivityAnalysis,
   ]);
-
-  const onAdjustTarget = (value, qtype) => {
-    const currentValue = currentSegmentDetail?.target
-      ? parseFloat(currentSegmentDetail.target)
-      : 0;
-    let newValue = {};
-    let adjustedTarget = 0;
-    if (qtype === "percentage" && percentageSensitivity) {
-      const absoluteValue = value ? (currentValue * value) / 100 : 0;
-      adjustedTarget = value ? (absoluteValue + currentValue).toFixed(2) : 0;
-      newValue = {
-        ...newValue,
-        [`${selectedSegment}_absolute-increase_adjusted-target`]:
-          parseFloat(adjustedTarget),
-        [`${selectedSegment}_percentage-increase_adjusted-target`]: value,
-      };
-    }
-    if (qtype === "absolute" && !percentageSensitivity) {
-      adjustedTarget = value || 0;
-      const absoluteChanged = value - currentValue;
-      const percentage = currentValue ? absoluteChanged / currentValue : 0;
-      const percentageIncrease = (percentage * 100).toFixed(2);
-      newValue = {
-        ...newValue,
-        [`${selectedSegment}_percentage-increase_adjusted-target`]:
-          parseFloat(percentageIncrease),
-        [`${selectedSegment}_absolute-increase_adjusted-target`]:
-          absoluteChanged,
-      };
-    }
-    newValue = {
-      ...newValue,
-      [`${selectedSegment}_adjusted-target`]: parseFloat(adjustedTarget),
-    };
-    setAdjustedValues((prev) => ({ ...prev, ...newValue }));
-  };
-
-  const handleOnSaveAdjustIncomeTarget = () => {
-    CaseVisualState.update((s) => ({
-      ...s,
-      sensitivityAnalysis: {
-        ...s.sensitivityAnalysis,
-        config: {
-          ...s.sensitivityAnalysis.config,
-          ...adjustedValues,
-        },
-      },
-    }));
-    setShowAdjustIncomeModal(false);
-  };
 
   const backFunction = useCallback(() => {
     navigate(`/case/${currentCase.id}/${stepPath.step3.label}`);
@@ -420,14 +312,6 @@ const AssessImpactMitigationStrategies = ({
     form.setFieldsValue(filteredValues);
   };
 
-  const onChangePercentage = (value) => {
-    if (value === "percentage") {
-      setPercentageSensitivity(true);
-    } else {
-      setPercentageSensitivity(false);
-    }
-  };
-
   return (
     <Row id="assess-impact-mitigation-strategies" gutter={[24, 24]}>
       {contextHolder}
@@ -556,171 +440,8 @@ const AssessImpactMitigationStrategies = ({
 
       {/* #3 Adjust Income Target */}
       <Col span={24}>
-        <Card className="card-section-wrapper adjust-income-target-wrapper">
-          <Row gutter={[20, 20]} align="middle">
-            <Col span={24}>
-              <div className="title">When do we reach the target?</div>
-            </Col>
-            <Col span={12}>
-              <div className="description">
-                Based on the selected drivers that are open to change, we run an
-                optimization model to search for the minimum required changes in
-                order to reach the income target.
-              </div>
-            </Col>
-            <Col span={12}>
-              <Space direction="vertical">
-                <div className="label">
-                  Select the (sub)drivers that you want to include in further
-                  analysis:
-                </div>
-                <AllDriverTreeSelector />
-              </Space>
-            </Col>
-            <Col span={12}>
-              <Button
-                className="button-ghost-white"
-                onClick={() => setShowAdjustIncomeModal(true)}
-              >
-                Adjust your income target
-              </Button>
-            </Col>
-            <Col span={12}>
-              <Space align="center">
-                <Switch />
-                <div className="label">Stay within feasible ranges?</div>
-              </Space>
-            </Col>
-          </Row>
-        </Card>
+        <AdjustIncomeTarget selectedSegment={selectedSegment} />
       </Col>
-
-      <Modal
-        title="Adjust your income target"
-        open={showAdjustIncomeModal}
-        onOk={handleOnSaveAdjustIncomeTarget}
-        okButtonProps={{
-          disabled: !enableEditCase,
-        }}
-        okText="Save income target"
-        onCancel={() => setShowAdjustIncomeModal(false)}
-        width="60%"
-        className="adjust-income-target-modal-container"
-        maskClosable={false}
-        cancelText={null}
-        footer={(_, { OkBtn }) => <OkBtn />}
-      >
-        <div className="description">
-          It might be interesting to change your income target after inspecting
-          the graphs below to observe how the results vary. If you do not adjust
-          the target, we will use the current target value for the calculations
-        </div>
-        <Row
-          align="middle"
-          gutter={[20, 20]}
-          className="adjust-income-target-step-wrapper"
-        >
-          <Col span={24}>
-            <Space className="step-wrapper" align="center">
-              <div className="number">1.</div>
-              <div className="label">Changing the target</div>
-            </Space>
-            <div className="description">
-              If you like to change the target, please choose whether you would
-              like to express the changes in current values using percentages or
-              absolute values.
-            </div>
-            <div className="description">
-              <Row>
-                <Col span={8}>
-                  <Select
-                    style={{ width: "100%" }}
-                    options={[
-                      {
-                        label: "Percentage",
-                        value: "percentage",
-                      },
-                      {
-                        label: "Absolute",
-                        value: "absolute",
-                      },
-                    ]}
-                    onChange={onChangePercentage}
-                    value={percentageSensitivity ? "percentage" : "absolute"}
-                  />
-                </Col>
-              </Row>
-            </div>
-          </Col>
-          <Col span={24}>
-            <Space className="step-wrapper" align="center">
-              <div className="number">2.</div>
-              <div className="label">Adjust current values below</div>
-            </Space>
-            <div className="description">
-              <Row gutter={[20, 20]} align="start">
-                <Col span={7}>
-                  <div className="input-label">Current Target</div>
-                  <div className="current-target-wrapper">
-                    {currentSegmentDetail
-                      ? thousandFormatter(
-                          currentSegmentDetail.target,
-                          determineDecimalRound(currentSegmentDetail.target)
-                        )
-                      : 0}{" "}
-                    <small>({currentCase?.currency})</small>
-                  </div>
-                </Col>
-                <Col span={10}>
-                  <div className="input-label">
-                    {percentageSensitivity ? "% Change" : "Adjusted Target"}
-                  </div>
-                  {["absolute", "percentage"].map((qtype) => (
-                    <div
-                      key={qtype}
-                      style={{
-                        display:
-                          qtype !== "percentage" && percentageSensitivity
-                            ? "none"
-                            : qtype === "percentage" && !percentageSensitivity
-                            ? "none"
-                            : "",
-                      }}
-                    >
-                      <InputNumber
-                        style={{
-                          width: "90%",
-                        }}
-                        addonAfter={qtype === "percentage" ? "%" : ""}
-                        {...InputNumberThousandFormatter}
-                        controls={false}
-                        onChange={(value) => onAdjustTarget(value, qtype)}
-                        value={
-                          percentageSensitivity
-                            ? adjustedValues?.[
-                                `${selectedSegment}_percentage-increase_adjusted-target`
-                              ]
-                            : adjustedValues?.[
-                                `${selectedSegment}_adjusted-target`
-                              ]
-                        }
-                      />
-                    </div>
-                  ))}
-                </Col>
-                <Col span={7}>
-                  <div className="input-label">
-                    {percentageSensitivity ? "Adjusted Target" : "% Change"}
-                  </div>
-                  <div className="adjusted-target-wrapper">
-                    {adustedTargetChange}
-                  </div>
-                </Col>
-              </Row>
-            </div>
-          </Col>
-        </Row>
-      </Modal>
       {/* EOL Adjust Income Target */}
     </Row>
   );
